feat: add basic Task class methods
Allow creation of new Task class instance, initializing with worker parameters and optionally with worker itself. Tasks with worker can be executed. Worker for any task can be modified utilizing the same set of worker parameters which cannot be modified once set. Result of the last execution of the task is stored for later retrieval.