feat: allow fallback config for execute method
A task must have a worker and workerParams defined at the time of execution. If any of them is undefined, then it should be set or provided via the fbConfig to use for the execution, else it will throw an error. fbConfig does not set the worker or workerParams but only provides the temporary value to fallback on. If even fallback values are undefined, then it throws an error.