feat: add GroupTask class
Allow creation of new GroupTask class instance, initializing with a type and an array of subTasks. Type of task represents the mode of execution that is series and parallel. In series, subTasks are executed one after another which can be used to order tasks based on order or dependency. In parallel, subTasks are executed simultaneously which can be used to groups taks which are independ of each other. Result of the execution is an array of results of each subTask.